Move helper functions for tiered tests into a new file
